Split OsClientCapabilities.
