+++ src/client/Microsoft.Identity.Client/OsCapabilitiesExtensionsPublic.cs
@@ -12,7 +12,7 @@ namespace Microsoft.Identity.Client
     /// <summary>
     /// Extension methods
     /// </summary>
-    public static class OsCapabilitiesExtensions // TODO MSAL5: move these to their parent types
+    public static class OsCapabilitiesExtensionsPublic // TODO MSAL5: move these to their parent types
     {
         /// <summary>
         /// Returns true if MSAL can use a system browser.
@@ -66,23 +66,6 @@ namespace Microsoft.Identity.Client
 
             throw new ArgumentException("This extension method is only available for the PublicClientApplication implementation " +
                 "of the IPublicClientApplication interface.");
-        }
-
-       
-
-        /// <summary>
-        /// Returns the certificate used to create this <see cref="ConfidentialClientApplication"/>, if any.
-        /// </summary>
-        public static X509Certificate2 GetCertificate(this IConfidentialClientApplication confidentialClientApplication)
-        {
-
-            if (confidentialClientApplication is ConfidentialClientApplication cca)
-            {
-                return cca.Certificate;
-            }
-
-            throw new ArgumentException("This extension method is only available for the ConfidentialClientApplication implementation " +
-                "of the IConfidentialClientApplication interface.");
         }
     }
 }
